fixes #91
