+++ package.json
@@ -34,6 +34,7 @@
   },
   "devDependencies": {
     "tap-spec": "^0.2.0",
-    "tape": "^2.13.3"
+    "tape": "^2.13.3",
+    "node-static": "~0.7.6"
   }
 }
